--- src/test/java/eu/yals/test/ui/elements/ClipboardHelperElement.java
@@ -1,14 +1,0 @@
-package eu.yals.test.ui.elements;
-
-import com.vaadin.testbench.TestBenchElement;
-import com.vaadin.testbench.elementsbase.Element;
-
-/**
- * A TestBench element representing a <code>&lt;clipboard-helper&gt;</code> element.
- *
- * @since 2.7
- */
-@Element("clipboard-helper")
-public class ClipboardHelperElement extends TestBenchElement {
-    // Empty class for further implementation
-}
