+++ src/test/java/eu/yals/test/ui/pageobjects/NotFoundViewPageObject.java
@@ -1,26 +1,15 @@
 package eu.yals.test.ui.pageobjects;
 
-import com.vaadin.flow.component.html.testbench.H1Element;
+import com.codeborne.selenide.SelenideElement;
 import eu.yals.ui.err.NotFoundView;
-import org.openqa.selenium.WebDriver;
+
+import static com.codeborne.selenide.Selenide.$;
 
 /**
  * Page Object for {@link NotFoundView}. Contains elements from NotFoundView
  *
  * @since 2.7
  */
-public class NotFoundViewPageObject extends YalsPageObject {
-
-    public static NotFoundViewPageObject getPageObject(WebDriver driver) {
-        return new NotFoundViewPageObject(driver);
-    }
-
-    public NotFoundViewPageObject(WebDriver driver) {
-        super(driver, NotFoundView.IDs.VIEW_ID);
-    }
-
-    // elements
-    public H1Element getTitle() {
-        return $(H1Element.class).first();
-    }
+public class NotFoundViewPageObject {
+    public static final SelenideElement TITLE = $("h1");
 }
